+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.4.0
+* ADD: isFileImage function was added for checking of files content-type
+* UPD: checkIsImageUrlAllowable function was renamed to isImageUrlAllowed and checks was extended
+
 ## 1.3.0
 * ADD: new module validationHelper and function checkIsImageUrlAllowable for checking of images URLs